requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
certifi
charset-normalizer==3.1.0
click==8.1.3
Django==4.1.7
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
idna==3.4
line-bot-sdk==2.4.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
yarl==1.8.2
